Fixed several clippy warnings
I went through all the clippy warnings and fixed the ones that made sense. A few of them I left alone for now because they are functions that might be used later so don't want to add the compiler flags just yet.

I also fixed everything that was mentioned in the PR comments.