refactor: back out from the `_full` functions in `Validator`
Those functions had too many unintuitive quirks in their API,
such as taking a reference to a closure and needing `&mut *sc` instead
of just `sc` in some cases.

Went back to triplets of functions: one for taking an `&mut sc`,
one for taking a `Scopes` value, and one for taking a closure.